Split typed expression parts
