+++ tests/docs_truth/repo_hygiene/ast_expression_operators.rs
@@ -26,3 +26,41 @@ fn ast_expression_operator_spelling_lives_in_focused_helper() {
         );
     }
 }
+
+#[test]
+fn typed_ast_expression_parts_live_in_focused_helper() {
+    let typed = read("src/ast/typed.rs");
+    let parts = read("src/ast/typed/expression_parts.rs");
+
+    for helper in [
+        "MatchKind",
+        "TypedMatchArm",
+        "TypedPattern",
+        "Capture",
+        "TypedStringPart",
+    ] {
+        assert!(
+            !typed.contains(&format!("pub struct {helper}"))
+                && !typed.contains(&format!("pub enum {helper}")),
+            "typed AST root should not own expression part helper: {helper}"
+        );
+        assert!(
+            parts.contains(&format!("pub struct {helper}"))
+                || parts.contains(&format!("pub enum {helper}")),
+            "typed AST expression part helper should live in focused helper: {helper}"
+        );
+    }
+
+    assert!(
+        typed.contains("mod expression_parts;"),
+        "typed AST root should include focused expression parts module"
+    );
+    assert!(
+        typed.contains("pub use expression_parts::{Capture, MatchKind, TypedMatchArm, TypedPattern, TypedStringPart};"),
+        "typed AST root should preserve public re-exports for expression parts"
+    );
+    assert!(
+        typed.lines().count() < 230,
+        "typed.rs should stay focused on typed expression, declaration, and program shells"
+    );
+}
